chore(tests): moved viewmodel tests to the test/viewmodel
